+++ molecularnodes/entities/__init__.py
@@ -1,4 +1,4 @@
 from .ensemble import CellPack, Ensemble, StarFile
 from .molecule.base import Molecule
-from .molecule.api.selection import Selector
+from .molecule.selection import Selector
 from .trajectory import OXDNA, Trajectory
